requirements.txt
# AWS Glue and PySpark dependencies
pyspark==3.1.1
boto3==1.26.137
awswrangler==2.19.0

# Apache Hudi
# Note: Hudi jars are typically provided by AWS Glue runtime

# Data processing
pandas==1.3.5
numpy==1.21.6
pyarrow==6.0.1

# Configuration
PyYAML==6.0

# Testing
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.10.0
moto==4.1.10

# Code quality
pylint==2.17.4
black==23.3.0
flake8==6.0.0

# Type checking
mypy==1.3.0